Backend2/requirements.txt
flask==3.0.2
flask-cors==4.0.0
PyMuPDF==1.23.26
requests==2.31.0
markdown==3.5.2
python-dotenv==1.0.1 